requirements.txt
pillow>=10.0.0
psutil>=5.9.0
packaging>=23.0
requests>=2.31.0
tomli>=2.0.1